[tnba-monitor] Make TNBA process similar to Triage (see commit message for details)
This commit refactors TNBA a little bit to make it closer to the structure of the Triage
process. The objective of this tiny refactor is to avoid future issues with concurrency,
to improve logging, and to make code easier to modify and understand since all tickets
are split into two subsets (tickets with TNBA and tickets without TNBA) before even
executing any logic.

Along with these improvements, this refactor prepares code for subsequent changes that
will allow taking into consideration multiple ticket details, instead of just one.