refactor(customer-cache): cross edges in stored and fresh caches attending to their serial numbers, instead of their EdgeIdentifier
When HA is enabled in an edge, this edge produces two entries in the cache of customers. Due to the fact that both entries are
related to the same EdgeIdentifier, the previous logic to cross stored and fresh caches could lead to scenarios where one of
the entries was missing in the final cache.

To avoid these cases, the crossing logic is now based on serial numbers since each device in a HA pair has a unique serial number.
Facing a collision between two devices with the same exact serial number is now very improbable.