cd(bruin-bridge): add Bruin DEV IPs to /etc/hosts at deployment time
Previously, these IPs were added to /etc/hosts while the Docker image of the service was built. This was working as expected
because ECS doesn't make changes to this file on its own, but Kubernetes does and indeed overwrites the entire file with
its own entries.

From now on, these IPs will be added to /etc/hosts at deployment time through a hostAliases section in the Deployment object
of the bruin-bridge subchart. That way Kubernetes will overwrite /etc/hosts contents as usual, but will also add entries to
have access to Bruin DEV from ephemeral environments.