fix(repair-tickets-monitor): fix metrics sending
The 'No validated service numbers' metric was lost
with the previous fix. This MR aims to send back
the lost metric.

The validated tickets (extracted tickets that exist
in Bruin) are now calculated before any work with
the service numbers, so that we can send the
validated ticket metrics in any case.
Additionally, the 'No validated service numbers'
metric is now explicitely calculated (instead of
reverting the exception raising that was removed
in the last fix).