feat(repair-tickets-monitor): add get_device_topics_rpc
Adds a new rpc folder in which to store rpcs.
The base classes for any rpc implementation are located
at rpc/base.
- Rpc: class that handles the integration with the EventBus
- RpcRequest: base class to be extended
- RpcResponse: convenience response wrapper class
- RpcLogger: LoggerAdapter class that prints the request_id
at each log

Also adds the get_device_topics_rpc as the first rpc
implementation.

Finally, adds the Device and Topic domain models.
These are used to build the get_device_topics_rpc API.

While the new RpcLogger is pretty simple, we should aim to
apply this LoggerAdapter strategy at the upper layers.
For that, we may need to build a "StackableLogger" that can
extend in some way the LoggerAdapter extra context in order
to print the upper layers context too.

The BruinRepository will start to be deprecated in favor
of the new rpc approach. Any maintenance done on any
BruinRepository method should try to migrate the method
to the new rpc model.