feat(repair-tickets-monitor): add new rpc component
Adds a new rpc folder in which to store rpcs.
The base classes for any rpc implementation are located
at base_rpc.
- Rpc: class that handles the integration with the EventBus
- RpcRequest: base class to be extended
- RpcResponse: convenience response wrapper class
- RpcLogger: LoggerAdapter class that prints the request_id
at each log

The BruinRepository will start to be deprecated in favor
of the new rpc approach. Any maintenance done on any
BruinRepository method should try to migrate the method
to the new rpc model.

While the new RpcLogger is pretty simple, we should aim to
apply this LoggerAdapter strategy at the upper layers.
For that, we may need to build a "StackableLogger" that can
extend in some way the LoggerAdapter extra context in order
to print the upper layers context too.