feat(bruin-bridge): add a new post email status action
The current mark email as done action doesn't support
any other status that "Done". RTA is going to need
to mark emails with status "New" or "AIQ".

Rather than updating the current mark email as done action,
a new one was built with the new components at hand.

It also makes use of the new SentenceFormatterService to
decouple this particular logic from the message handling
code.